Sprint1 into Stage For Deployment!! (#42)
* Trying4 (#37)

* moved design document into doc folder

* updated for all changes

* added seatMapWidget files

* Created component & corresponding routing for the CSXLSeatMap Component

* Changed title of CSXLSeatMapComponent

* Working on SVG Graphics for CSXL Seat Map

* changes

* trying to save progress

* updated confirming

* coding

* backend seats working

* backend fully finished

---------


* Backend Test and Frontend Ambassador Modifications (#39)

* Add special hour blocks for furniture install (#352)

* Fix low-hanging vulnerabilities (#354)

* Fix conflicting room reservations bug (#355)

* Prevent double booking of the same room

The reservation draft API does not prevent a user from double booking a
room some other user has already booked. While (generally) not possible
via the frontend, this ensures the backend does not permit double booking.

Tests added for various relevant cases.

* Fix failing tests for reservations and warning in organizations

* Fix front-end issue with draft drop-in reservation (#356)

Due to some changes that merged with room reservations, when creating a
drop-in reservation and navigating back (intentionally or not) to home
you would not see the draft reservation anymore because it was filtered
from the home page. This meant a draft reservation was created but could
not be cancelled by user or ambassador. This PR fixes this issue by
making the draft of a drop-in visible on the home page. No changes in
behavior for room reservations.

* Upgrade FastAPI for DDoS security fix to 0.110.0 (#358)

* Profile Page Improvements (#359)

* Created a profile module to contain the new Profile Page and Profile Editor.
* Created a new Profile Page component and associated HTML and CSS files, which includes a button route to the Profile Editor.
* Created a new User Profile Card widget to display a users name, profile picture, email, and pronouns.
* Moved Github Link and Bearer Token to Profile Page and made minor changes to the styling.
* Created a "Profile Actions" card for github link, community agreement, and auth bearer token.

* Fix profile card styling (#361)

* Update policy for furniture install (#362)

* Prevent communal area drop-ins while furniture installation completes (#363)

* Update policy for furniture install

* Block off communal area for OH

* moved design document into doc folder

* updated for all changes

* Make profile page responsive (#366)

* 374 Fix active reservations boundary logic (#375)

* Update filters to show active reservations on ambassador view

* Fix bug when start_idx < 0

* Only consider rooms that are present in map

---------


* added seatMapWidget files

* Created component & corresponding routing for the CSXLSeatMap Component

* Changed title of CSXLSeatMapComponent

* Working on SVG Graphics for CSXL Seat Map

* Reset policies for Phase 2 Launch of XL (#376)

* Fix broken build in prod due to npm memory limit (#377)

* Fix @angular-devkit/build-angular dependency vulnerabilities in production (#378)

* Fix vulnerability with @angular-devkit/build-angular dep

* Try reducing memory to match cloudapps pod limit (#379)

* Increase memory limit for deployment (#381)

* Hot fixing office hours schedule (#382)

* Fix bug with end_idx logic for future reservations (#384)

* Fix bug with end_idx logic for future reservations

The current problem with the code is that at 11 pm, we were
not able to see future reservations in red for tomorrow. This was
actually a red herring that led me to look at parts of the code
that were not responsible for this.

Then suddenly at 12 am, everything started working as intended.
This was the biggest hint for me in debugging this.

Basically on line 345, I just had to add in a condition to check
that the date I am getting as a param is the same date as today.
Only in that case, should I execute the logic of
end_idx < current_time_idx. However, earlier this was being
executed for all reservations which is why whenever we had
any reservations which had an end time of less than 11 pm
(which most reservations do), our function would just skip over
them. This also aligns with what we saw at 12 am when
everything seemed to work as intended. This is again because
most reservations don't have an ending time before 12 am.

Anyway, took me over 60 minutes to add in these 30 chars to fix
the bug. So I typed an average of 0.5 char per minute... :,)

* Added kind of a test for future reservations

Although this test is time dependent, and
we will need some way of freezing time
to accurately test the efficacy of the code.

* changes

* Change Dockerfile to use yarn instead of npm (#393)

* trying to save progress

* updated confirming

* coding

* backend seats working

* 399 add checkin time to avoid cancellation verbiage (#400)

* Add checkin deadline verbiage

* Add parameter to checkinDeadline

Since users can reserve a reservation up until the end a reservation time period, the checkinDeadline was exceeding the reservation's end time.
To correct this take the minimum of the start time + 10 mins and the end time.

* backend fully finished

* Added front-end service and connected with seat map widget to gain room data from backend database and API

* Seats can not be selected when reservable is false, and front-end rectangular and circular tables were updated from the backend

* Finished creating backend services for RoomService methods that were created

* Functional code for Walkin Ambassador Reservations

* Updated to push to stage

---------


* All Integrated Sprint 1 (#40)

* moved design document into doc folder

* updated for all changes

* added seatMapWidget files

* Created component & corresponding routing for the CSXLSeatMap Component

* Changed title of CSXLSeatMapComponent

* Working on SVG Graphics for CSXL Seat Map

* changes

* trying to save progress

* updated confirming

* coding

* backend seats working

* backend fully finished

* Added front-end service and connected with seat map widget to gain room data from backend database and API

* Seats can not be selected when reservable is false, and front-end rectangular and circular tables were updated from the backend

* Finished creating backend services for RoomService methods that were created

* Adding frontend fixes to clicked seats to make sure it limits to 4

* everything integrated both views as well asmaking sure ambassador view displays the multiple seats

---------


* Update A7-spec.md

Added descriptions about the technical and user experience designs and the development concerns.

* Add seat demo data (#41)

* moved design document into doc folder

* updated for all changes

* added seatMapWidget files

* Created component & corresponding routing for the CSXLSeatMap Component

* Changed title of CSXLSeatMapComponent

* Working on SVG Graphics for CSXL Seat Map

* changes

* trying to save progress

* updated confirming

* coding

* backend seats working

* backend fully finished

* Added front-end service and connected with seat map widget to gain room data from backend database and API

* Seats can not be selected when reservable is false, and front-end rectangular and circular tables were updated from the backend

* Finished creating backend services for RoomService methods that were created

* Added seat demo data realistic to the CSXL floorplan

---------


---------

+++ frontend/src/app/coworking/widgets/seat-map/seat-map.widget.ts
@@ -1,8 +1,14 @@
-import { Component, OnInit } from '@angular/core';
-import { Seat, SeatAvailability } from 'src/app/coworking/coworking.models';
+import { Component, EventEmitter, OnInit, Output } from '@angular/core';
+import {
+  CircleTable,
+  RectangleTable,
+  Seat,
+  SeatAvailability
+} from 'src/app/coworking/coworking.models';
 import { CsxlSeatMapService } from '../../seating-reservation/csxl-seat-map/csxl-seat-map.service';
 import { CoworkingService } from '../../coworking.service';
 import { ReservationService } from '../../reservation/reservation.service';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'seat-map',
@@ -12,25 +18,73 @@ import { ReservationService } from '../../reservation/reservation.service';
 export class SeatMapWidgetComponent implements OnInit {
   seats: Seat[] = [];
   seatsClicked: Seat[] = [];
-  seatsReserved: Seat[] = []; // seats only become reserved once user presses reserved button
+  seatsReserved: SeatAvailability[] = []; // seats only become reserved once user presses reserved button
+  @Output() reservationDrafted = new EventEmitter<SeatAvailability[]>();
+  boundaries: String = '';
+  rectangleTables: RectangleTable[] = [];
+  circleTables: CircleTable[] = [];
+
+  roomId = 'SN156';
 
   constructor(
-    private mapService: CsxlSeatMapService,
+    public mapService: CsxlSeatMapService,
     private coworkService: CoworkingService,
-    private reserveService: ReservationService
+    private reserveService: ReservationService,
+    private snackBar: MatSnackBar
   ) {}
   // seats only become reserved once user presses reserved button
 
   ngOnInit(): void {
-    this.seats = this.mapService.getSeats();
+    this.getSeats();
+    this.getBoundaries(this.roomId);
+    this.getCircleTables(this.roomId);
+    this.getRectangleTables(this.roomId);
+    this.getBoundaries(this.roomId);
+    this.getCircleTables(this.roomId);
+    this.getRectangleTables(this.roomId);
+  }
+
+  getSeats(): void {
+    this.mapService.getSeats(this.roomId).subscribe((seats) => {
+      this.seats = seats;
+    });
   }
 
   seatClicked(seat: Seat) {
-    if (this.mapService.isSeatClicked(seat)) {
-      this.mapService.removeClickedSeat(seat);
+    if (this.isReservable(seat)) {
+      if (this.mapService.isSeatClicked(seat)) {
+        this.mapService.removeClickedSeat(seat);
+      } else {
+        this.mapService.addClickedSeat(seat);
+      }
     } else {
-      this.mapService.addClickedSeat(seat);
+      this.snackBar.open('This seat is already reserved', 'Close', {
+        duration: 3000
+      });
     }
+  }
+
+  getBoundaries(id: string) {
+    this.mapService.getBoundaries(id).subscribe((boundaries) => {
+      this.boundaries = boundaries;
+    });
+  }
+
+  getCircleTables(id: string) {
+    this.mapService.getCircleTables(id).subscribe((circleTables) => {
+      this.circleTables = circleTables;
+    });
+    console.log(this.circleTables);
+  }
+
+  getRectangleTables(id: string) {
+    this.mapService.getRectangleTables(id).subscribe((rectangleTables) => {
+      this.rectangleTables = rectangleTables;
+    });
+  }
+
+  isReservable(seat: Seat) {
+    return seat.reservable;
   }
 
   isButtonActive(): boolean {
@@ -42,13 +96,20 @@ export class SeatMapWidgetComponent implements OnInit {
   }
 
   reserveSeats() {
-    this.seatsReserved = this.mapService.getClickedSeats();
-    this.mapService.reserveSeats(this.seatsReserved);
-    let reservedSeatAvails: SeatAvailability[] =
-      this.mapService.convertSeatsToSeatAvailability(this.seatsReserved);
-    this.coworkService.draftReservation(reservedSeatAvails).subscribe({
-      next: (value) => this.reserveService.confirm(value)
-    });
+    const clickedSeats = this.mapService.getClickedSeats();
+    const reservations: SeatAvailability[] = clickedSeats.map((seat) => ({
+      id: seat.id,
+      availability: seat.availability,
+      title: seat.title,
+      shorthand: seat.shorthand,
+      reservable: seat.reservable,
+      has_monitor: seat.has_monitor,
+      sit_stand: seat.sit_stand,
+      x: seat.x,
+      y: seat.y
+    }));
+    this.reservationDrafted.emit(reservations);
+    this.mapService.clearReservations();
   }
 
   clearReserves() {
